Extract loading the card catalog into its own file so we can easily mock it during unit testing of carddb.
